WIP: Outline for collector pattern for grub boot items
/cc @acaiafa **DO NOT MERGE**

Started working on a rough outline on how to configure the individual
items using the collector pattern. Key bit here is the `include
Poise(container: true)` bit. Then each of the `grub_item`s can declare
grub as the parent resource and it can collect their `#to_s` output and
do whatever it needs to do.